+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Rong 2019 and van Hest 2005 mycophenolic acid models, as re-implemented by Maizaud 2025 ([doi:10.1371/journal.pone.0330854](https://doi.org/10.1371/journal.pone.0330854)) - adult kidney transplant recipients on tacrolimus (Rong) or ciclosporin (van Hest).
+- Add Poels 2025 elranatamab QSP model ([doi:10.1038/s41540-025-00585-z](https://doi.org/10.1038/s41540-025-00585-z)) -- adults with relapsed or refractory multiple myeloma.
 
 - Add Moein 2024 apitolisib translational PK/PD models ([doi:10.1007/s40268-024-00459-5](https://doi.org/10.1007/s40268-024-00459-5)) -- mouse (786-O renal cell adenocarcinoma xenograft) and adults with advanced solid tumors or non-Hodgkin's lymphoma.